# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Auth API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/community/messages": {
            "get": {
                "tags": [
                    "Community Messages"
                ],
                "summary": "Get all community messages",
                "operationId": "76d777b252037b63d397620a7a484fa8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pinned",
                        "in": "query",
                        "description": "Show only pinned messages",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "announcement",
                        "in": "query",
                        "description": "Show only announcements",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "last_updated",
                        "in": "query",
                        "description": "Last update timestamp for polling",
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Messages retrieved successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Community Messages"
                ],
                "summary": "Create a new community message",
                "operationId": "01b3d6cec5c7431b25e76dc57faac774",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Message title",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Message content",
                                        "type": "string"
                                    },
                                    "category": {
                                        "description": "Message category",
                                        "type": "string"
                                    },
                                    "tags": {
                                        "description": "Message tags (comma-separated string or array)",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "is_pinned": {
                                        "description": "Pin the message (true/false, 1/0, yes/no)",
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "is_announcement": {
                                        "description": "Mark as announcement (true/false, 1/0, yes/no)",
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "attachments[]": {
                                        "description": "File attachments",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/community/messages/{id}": {
            "get": {
                "tags": [
                    "Community Messages"
                ],
                "summary": "Get a specific community message",
                "operationId": "c45bac022142608378085247fcdd24f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message retrieved successfully"
                    },
                    "404": {
                        "description": "Message not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Community Messages"
                ],
                "summary": "Update a community message",
                "operationId": "4c7b7231e9cdf13337229a3906ecd631",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Message title",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Message content",
                                        "type": "string"
                                    },
                                    "category": {
                                        "description": "Message category",
                                        "type": "string"
                                    },
                                    "tags": {
                                        "description": "Message tags (comma-separated string or array)",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "is_pinned": {
                                        "description": "Pin the message (true/false, 1/0, yes/no)",
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "is_announcement": {
                                        "description": "Mark as announcement (true/false, 1/0, yes/no)",
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "attachments[]": {
                                        "description": "File attachments",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message updated successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Message not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Community Messages"
                ],
                "summary": "Delete a community message",
                "operationId": "3a88ec8ea8f5b253d46c265b0ac50d6c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Message not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/community/messages/{id}/like": {
            "post": {
                "tags": [
                    "Community Messages"
                ],
                "summary": "Toggle like on a community message",
                "operationId": "1d90b2774c71d980245e9b64231e4864",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Like toggled successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/community/messages/latest": {
            "get": {
                "tags": [
                    "Community Messages"
                ],
                "summary": "Get latest messages for polling",
                "operationId": "34763d75687a590d55697019c579bb98",
                "parameters": [
                    {
                        "name": "last_id",
                        "in": "query",
                        "description": "Last message ID seen",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Latest messages retrieved"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}